requirements.txt
streamlit==1.28.0
psycopg2-binary==2.9.7
redis==4.6.0
pandas==2.0.3
plotly==5.15.0
kafka-python==2.0.2
pyspark==3.4.1
faker==19.3.0
psutil==5.9.5